Added filter to prevent file being deleted if desired (defaults to true, will delete file).

+++ src/Jobs/CreateMerchantReportedConversionReport.php
@@ -239,7 +239,9 @@ class CreateMerchantReportedConversionReport extends AbstractBatchedActionSchedu
 			if ( $results['success'] ) {
 				// Delete CSV files.
 				foreach ( $file_paths as $file_path ) {
-					$this->writer->delete_file( $file_path );
+					if ( apply_filters( 'woocommerce_gla_youtube_orders_csv_delete_on_complete', true ) ) {
+						$this->writer->delete_file( $file_path );
+					}
 				}
 
 				// Remove file state for this date.
